+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`142` Allow registering plugin methods on custom hooks
 * :bug:`143` Use gossip's inernal handler exception hook to debug hook failures when ``--pdb`` is used
 * :release:`0.6.0 <21-05-2014>`
 * :feature:`-` Added assertion introspection via AST rewrite, borrowed from `pytest <http://pytest.org>`_.